// src/models/IFormData.ts
export default interface IFormData {
  nome: string
  email: string
  telefoneComercial: string
  tipoCulinaria: string
  cnpj: string
  tempoPreparo: string
  horarioFuncionamento: string
  taxaEntrega: string
  faixaPreco: string
  cep: string
  cidade: string
  rua: string
  numero: string
  pontoReferencia: string
}